Adds example for `fixedFirstColumn` feature in `DataTable`
