+++ polaris-react/playground/Playground.tsx
@@ -1,359 +1,11 @@
-import React, {useState, useCallback} from 'react';
+import React from 'react';
 
-import {Card, Link, Page, DataTable} from '../src';
-
-function FullDataTableExample() {
-  const [sortedRows, setSortedRows] = useState(null);
-
-  const initiallySortedRows = [
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="emerald-silk-gown"
-      >
-        Emerald Silk Gown
-      </Link>,
-      '$875.00',
-      124689,
-      140,
-      '$121,500.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="mauve-cashmere-scarf"
-      >
-        Mauve Cashmere Scarf
-      </Link>,
-      '$230.00',
-      124533,
-      83,
-      '$19,090.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="navy-merino-wool"
-      >
-        Navy Merino Wool Blazer with khaki chinos and yellow belt
-      </Link>,
-      '$445.00',
-      124518,
-      32,
-      '$14,240.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="emerald-silk-gown"
-      >
-        Emerald Silk Gown
-      </Link>,
-      '$875.00',
-      124689,
-      140,
-      '$121,500.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="mauve-cashmere-scarf"
-      >
-        Mauve Cashmere Scarf
-      </Link>,
-      '$230.00',
-      124533,
-      83,
-      '$19,090.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="navy-merino-wool"
-      >
-        Navy Merino Wool Blazer with khaki chinos and yellow belt
-      </Link>,
-      '$445.00',
-      124518,
-      32,
-      '$14,240.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="emerald-silk-gown"
-      >
-        Emerald Silk Gown
-      </Link>,
-      '$875.00',
-      124689,
-      140,
-      '$121,500.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="mauve-cashmere-scarf"
-      >
-        Mauve Cashmere Scarf
-      </Link>,
-      '$230.00',
-      124533,
-      83,
-      '$19,090.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="navy-merino-wool"
-      >
-        Navy Merino Wool Blazer with khaki chinos and yellow belt
-      </Link>,
-      '$445.00',
-      124518,
-      32,
-      '$14,240.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="emerald-silk-gown"
-      >
-        Emerald Silk Gown
-      </Link>,
-      '$875.00',
-      124689,
-      140,
-      '$121,500.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="mauve-cashmere-scarf"
-      >
-        Mauve Cashmere Scarf
-      </Link>,
-      '$230.00',
-      124533,
-      83,
-      '$19,090.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="navy-merino-wool"
-      >
-        Navy Merino Wool Blazer with khaki chinos and yellow belt
-      </Link>,
-      '$445.00',
-      124518,
-      32,
-      '$14,240.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="emerald-silk-gown"
-      >
-        Emerald Silk Gown
-      </Link>,
-      '$875.00',
-      124689,
-      140,
-      '$121,500.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="mauve-cashmere-scarf"
-      >
-        Mauve Cashmere Scarf
-      </Link>,
-      '$230.00',
-      124533,
-      83,
-      '$19,090.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="navy-merino-wool"
-      >
-        Navy Merino Wool Blazer with khaki chinos and yellow belt
-      </Link>,
-      '$445.00',
-      124518,
-      32,
-      '$14,240.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="emerald-silk-gown"
-      >
-        Emerald Silk Gown
-      </Link>,
-      '$875.00',
-      124689,
-      140,
-      '$121,500.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="mauve-cashmere-scarf"
-      >
-        Mauve Cashmere Scarf
-      </Link>,
-      '$230.00',
-      124533,
-      83,
-      '$19,090.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-    [
-      <Link
-        removeUnderline
-        url="https://www.example.com"
-        key="navy-merino-wool"
-      >
-        Navy Merino Wool Blazer with khaki chinos and yellow belt
-      </Link>,
-      '$445.00',
-      124518,
-      32,
-      '$14,240.00',
-      124689,
-      140,
-      '$121,500.00',
-    ],
-  ];
-
-  const rows = sortedRows ? sortedRows : initiallySortedRows;
-  const handleSort = useCallback(
-    (index, direction) => setSortedRows(sortCurrency(rows, index, direction)),
-    [rows],
-  );
-
-  return (
-    <Page title="Sales by product">
-      <Card>
-        <DataTable
-          columnContentTypes={[
-            'text',
-            'numeric',
-            'numeric',
-            'numeric',
-            'numeric',
-            'numeric',
-            'numeric',
-            'numeric',
-            'numeric',
-            'numeric',
-            'numeric',
-          ]}
-          headings={[
-            'Product',
-            'Price',
-            'SKU Number',
-            'Net quantity',
-            'Net sales',
-            'SKU Number',
-            'Net quantity',
-            'Net sales',
-          ]}
-          rows={rows}
-          totals={['', '', '', 255, '$155,830.00', '', 255, '$155,830.00']}
-          sortable={[true, true, false, false, true]}
-          defaultSortDirection="descending"
-          initialSortColumnIndex={4}
-          onSort={handleSort}
-          // footerContent={`Showing ${rows.length} of ${rows.length} results`}
-          stickyHeader
-          hasFixedFirstColumn
-          truncate
-        />
-      </Card>
-    </Page>
-  );
-
-  function sortCurrency(rows, index, direction) {
-    return [...rows].sort((rowA, rowB) => {
-      const amountA = parseFloat(rowA[index].substring(1));
-      const amountB = parseFloat(rowB[index].substring(1));
-
-      return direction === 'descending' ? amountB - amountA : amountA - amountB;
-    });
-  }
-}
+import {Page} from '../src';
 
 export function Playground() {
-  return <FullDataTableExample />;
+  return (
+    <Page title="Playground">
+      {/* Add the code you want to test in here */}
+    </Page>
+  );
 }
